requirements.txt
python-dotenv==0.19.2
requests==2.27.1
pytest==6.2.5
pytest-cov==2.12.1
responses==0.18.0
